Move `random.js` from `src/sidebar/util` to `src/shared`
This file will be useful for components in `src/annotator` so it will be
better to move it to common path for both the sidebar and annotator.
