+++ src/shared/random.js
