+++ src/sidebar/util/postmessage-json-rpc.js
@@ -1,4 +1,4 @@
-import { generateHexString } from './random';
+import { generateHexString } from '../../shared/random';
 
 /** Generate a random ID to associate RPC requests and responses. */
 function generateId() {
